requirements.txt
git+https://github.com/Rapptz/discord-ext-menus
PyNaCl==1.3.0
pandas
discord.py
apscheduler
aiosqlite
aiohttp
dnspython==1.16.0
async-timeout==3.0.1